Fix transient requires generates undeterministic test discovery
A require of a namespace that requires all cards were being used to
centralize cards discovery for the cards ui and the node tests.
Node test in watch mode loses track of the correct dependency on a file
state change if the requires are transiente.
Changed source code to require cards in the entrypoint namespace.