Use localstorage instead of sessionstorage to save mock state
While the session storage offers great flexibility, allowing two
different tabs of the same page to have different mock states, in
realitiy the developer experience is suboptimal. That's due to the
fact that I tend to close and open a single page many times, and would
like the mock state to be kept between multiple tab instances.